+++ select_test.go
@@ -1,245 +1,140 @@
 package sqlb
 
 import (
+    "fmt"
     "testing"
 
     "github.com/stretchr/testify/assert"
 )
 
-type selClauseTest struct {
-    c *selectClause
+type selectQueryTest struct {
+    q *SelectQuery
     qs string
     qargs []interface{}
+    qe error
 }
 
-func TestSelectClause(t *testing.T) {
+func TestSelectQuery(t *testing.T) {
     assert := assert.New(t)
 
     m := testFixtureMeta()
-    users := m.TableDef("users")
+    users := m.Table("users")
     articles := m.TableDef("articles")
     article_states := m.TableDef("article_states")
-    colUserName := users.ColumnDef("name")
-    colUserId := users.ColumnDef("id")
-    colArticleId := articles.ColumnDef("id")
-    colArticleAuthor := articles.ColumnDef("author")
+    colUserId := users.Column("id")
+    colUserName := users.Column("name")
+    colArticleId := articles.Column("id")
+    colArticleAuthor := articles.Column("author")
     colArticleState := articles.ColumnDef("state")
     colArticleStateId := article_states.ColumnDef("id")
     colArticleStateName := article_states.ColumnDef("name")
 
-    tests := []selClauseTest{
-        // a literal value
-        selClauseTest{
-            c: &selectClause{
-                projections: []projection{&value{val: 1}},
-            },
-            qs: "SELECT ?",
-            qargs: []interface{}{1},
-        },
-        // a literal value aliased
-        selClauseTest{
-            c: &selectClause{
-                projections: []projection{
-                    &value{alias: "foo", val: 1},
-                },
-            },
-            qs: "SELECT ? AS foo",
-            qargs: []interface{}{1},
-        },
-        // two literal values
-        selClauseTest{
-            c: &selectClause{
-                projections: []projection{
-                    &value{val: 1},
-                    &value{val: 1},
-                },
-            },
-            qs: "SELECT ?, ?",
-            qargs: []interface{}{1, 2},
-        },
-        // TableDef and ColumnDef
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName},
-            },
-            qs: "SELECT users.name FROM users",
-        },
-        // Table and ColumnDef
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users.Table()},
-                projections: []projection{colUserName},
-            },
-            qs: "SELECT users.name FROM users",
-        },
-        // TableDef and Column
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName.Column()},
-            },
-            qs: "SELECT users.name FROM users",
-        },
-        // aliased Table and Column
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users.As("u")},
-                projections: []projection{
-                    users.As("u").Column("name"),
-                },
-            },
-            qs: "SELECT u.name FROM users AS u",
-        },
-        // TableDef and multiple ColumnDef
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserId, colUserName},
-            },
-            qs: "SELECT users.id, users.name FROM users",
-        },
-        // TableDef and mixed Column and ColumnDef
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserId, colUserName.Column()},
-            },
+    tests := []selectQueryTest{
+        // Simple FROM
+        selectQueryTest{
+            q: Select(users),
             qs: "SELECT users.id, users.name FROM users",
         },
         // Simple WHERE
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName},
-                where: &whereClause{
-                    filters: []*Expression{
-                        Equal(colUserName, "foo"),
-                    },
-                },
-            },
-            qs: "SELECT users.name FROM users WHERE users.name = ?",
+        selectQueryTest{
+            q: Select(users).Where(Equal(colUserName, "foo")),
+            qs: "SELECT users.id, users.name FROM users WHERE users.name = ?",
             qargs: []interface{}{"foo"},
         },
-        // Simple LIMIT
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName},
-                limit: &limitClause{limit: 10},
-            },
-            qs: "SELECT users.name FROM users LIMIT ?",
-            qargs: []interface{}{10},
+        // Simple GROUP BY
+        selectQueryTest{
+            q: Select(users).GroupBy(colUserName),
+            qs: "SELECT users.id, users.name FROM users GROUP BY users.name",
         },
         // Simple ORDER BY
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName},
-                orderBy: &orderByClause{
-                    scols: []*sortColumn{colUserName.Desc()},
-                },
-            },
-            qs: "SELECT users.name FROM users ORDER BY users.name DESC",
+        selectQueryTest{
+            q: Select(users).OrderBy(colUserName.Desc()),
+            qs: "SELECT users.id, users.name FROM users ORDER BY users.name DESC",
         },
-        // Simple GROUP BY
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName},
-                groupBy: &groupByClause{
-                    cols: []projection{colUserName},
-                },
-            },
-            qs: "SELECT users.name FROM users GROUP BY users.name",
-        },
-        // GROUP BY, ORDER BY and LIMIT
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{colUserName},
-                groupBy: &groupByClause{
-                    cols: []projection{colUserName},
-                },
-                orderBy: &orderByClause{
-                    scols: []*sortColumn{colUserName.Desc()},
-                },
-                limit: &limitClause{limit: 10},
-            },
-            qs: "SELECT users.name FROM users GROUP BY users.name ORDER BY users.name DESC LIMIT ?",
+        // Simple LIMIT
+        selectQueryTest{
+            q: Select(users).Limit(10),
+            qs: "SELECT users.id, users.name FROM users LIMIT ?",
             qargs: []interface{}{10},
         },
-        // Simple sub-SELECT
-        selClauseTest{
-            c: &selectClause{
-                alias: "u",
-                selections: []selection{users},
-                projections: []projection{colUserName},
-            },
-            qs: "(SELECT users.name FROM users) AS u",
+        // Simple LIMIT with OFFSET
+        selectQueryTest{
+            q: Select(users).LimitWithOffset(10, 20),
+            qs: "SELECT users.id, users.name FROM users LIMIT ? OFFSET ?",
+            qargs: []interface{}{10, 20},
         },
-        // Single join
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{articles},
-                projections: []projection{colArticleId, colUserName.As("author")},
-                joins: []*joinClause{
-                    &joinClause{
-                        left: articles,
-                        right: users,
-                        onExprs: []*Expression{
-                            Equal(colArticleAuthor, colUserId),
-                        },
-                    },
-                },
-            },
+        // Simple sub-SELECT
+        selectQueryTest{
+            q: Select(users).As("u"),
+            qs: "(SELECT users.id, users.name FROM users) AS u",
+        },
+        // Simple INNER JOIN
+        selectQueryTest{
+            q: Select(colArticleId, colUserName.As("author")).Join(users, Equal(colArticleAuthor, colUserId)),
             qs: "SELECT articles.id, users.name AS author FROM articles JOIN users ON articles.author = users.id",
         },
-        // Multiple joins
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{articles},
-                projections: []projection{colArticleId, colUserName.As("author"), colArticleStateName.As("state")},
-                joins: []*joinClause{
-                    &joinClause{
-                        left: articles,
-                        right: users,
-                        onExprs: []*Expression{
-                            Equal(colArticleAuthor, colUserId),
-                        },
-                    },
-                    &joinClause{
-                        left: articles,
-                        right: article_states,
-                        onExprs: []*Expression{
-                            Equal(colArticleState, colArticleStateId),
-                        },
-                    },
-                },
-            },
+        // Two JOINs using Join() method
+        selectQueryTest{
+            q: Select(
+                colArticleId,
+                colUserName.As("author"),
+                colArticleStateName.As("state"),
+            ).Join(users, Equal(colArticleAuthor, colUserId),
+            ).Join(article_states, Equal(colArticleState, colArticleStateId)),
             qs: "SELECT articles.id, users.name AS author, article_states.name AS state FROM articles JOIN users ON articles.author = users.id JOIN article_states ON articles.state = article_states.id",
-        },
-        // Simple COUNT(*) on a table
-        selClauseTest{
-            c: &selectClause{
-                selections: []selection{users},
-                projections: []projection{Count()},
-            },
-            qs: "SELECT COUNT(*) FROM users",
         },
     }
     for _, test := range tests {
-        expLen := len(test.qs)
-        s := test.c.size()
-        assert.Equal(expLen, s)
-
-        expArgc := len(test.qargs)
-        assert.Equal(expArgc, test.c.argCount())
-
-        b := make([]byte, s)
-        written, _ := test.c.scan(b, test.qargs)
-
-        assert.Equal(written, s)
-        assert.Equal(test.qs, string(b))
+        if test.qe != nil {
+            assert.Equal(test.qe, test.q.Error())
+            continue
+        } else if test.q.Error() != nil {
+            qe := test.q.Error()
+            assert.Fail(qe.Error())
+            continue
+        }
+        qs, qargs := test.q.StringArgs()
+        assert.Equal(len(test.qargs), len(qargs))
+        assert.Equal(test.qs, qs)
     }
+}
+
+func TestModifyingSelectQueryUpdatesBuffer(t *testing.T) {
+    assert := assert.New(t)
+
+    m := testFixtureMeta()
+    users := m.TableDef("users")
+
+    q := Select(users)
+
+    qs, qargs := q.StringArgs()
+    assert.Equal("SELECT users.id, users.name FROM users", qs)
+    assert.Nil(qargs)
+
+    // Modify the underlying SELECT and verify string and args changed
+    q.Where(Equal(users.Column("id"), 1))
+    qs, qargs = q.StringArgs()
+    assert.Equal("SELECT users.id, users.name FROM users WHERE users.id = ?", qs)
+    assert.Equal([]interface{}{1}, qargs)
+}
+
+func TestSelectQueryErrors(t *testing.T) {
+    assert := assert.New(t)
+
+    q := &SelectQuery{}
+
+    assert.False(q.IsValid()) // Doesn't have a selectClause yet...
+    assert.Nil(q.Error()) // But there is no error set yet...
+
+    m := testFixtureMeta()
+    users := m.TableDef("users")
+
+    q = Select(users)
+
+    assert.True(q.IsValid())
+    assert.Nil(q.Error())
+
+    q.e = fmt.Errorf("Cannot determine left side of JOIN expression.")
+    assert.False(q.IsValid())
+    assert.NotNil(q.Error())
 }
